requirements.txt
requests>=2.28.0
python-dotenv>=0.21.0
openai>=1.0.0
argparse; python_version<'3.7' 